Try to avoid commits halfway through block writes
When a write request only affects the first part of a block, the block
is added to the transaction and future writes will just append to it
without repeating the CoW. Once the transaction ends, though, any blocks
that are still not complete will be subject to CoW if more data is
added. This is a problem because of the fragmentation it introduces into
large files that could otherwise be stored as a single extent.

As an (apparently successful) attempt to prevent this, check if the
block was written in full before calling apfs_transaction_commit() from
apfs_write_end(), and if so use slightly different thresholds to decide
if a commit is needed. The intended result is that the transaction will
go on a little longer to give a chance for the write to complete.

For the record, not much thought has gone into picking the new
thresholds.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>