Don't read dstream xattr blocks one by one
Currently apfs_xattr_extents_read() loops once through all the blocks in
the dstream, reading each of them from disk and copying it to the
destination buffer right away. It's much more effective to do IO for
several blocks at once, similar to the readahead used by regular file
dstreams.

So use two loops instead: one to dispatch the read requests, and a
second to wait for each of them and copy the results to the buffer.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>