Always check for failure after apfs_alloc_query()
There are two places in the code where I've neglected to check the
return value of apfs_alloc_query(). One is in apfs_btree_query(): this
was one of the first things I implemented years ago, so the mistake was
just inexperience. The second is in apfs_query_set_before_first(), a
more recent addition: I probably copied the code from apfs_btree_query()
withut a proper review.

Anyway, this would cause npd if apfs_alloc_query() failed, so add the
proper checks and cleanups.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>