Avoid overflow on sm_ip_block_count corruption
The sm_ip_block_count field of the on-disk spaceman has an apparently
excessive width of 64 bits. After reading it, apfs_ip_find_free() casts
it to an int, which probably makes sense for valid values, but not for a
corrupted filesystem. This doesn't seem to trigger an out-of-bounds read
because sb->s_blocksize is unsigned, but it looks awkward, so fix it.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>