Avoid fake buffer head on node splits
Rework apfs_node_split() to take advantage of the apfs_object changes
from last patch and avoid the fake buffer head. Now the temporary node
used for the split can be handled like any other node, so do that
instead of using big stack allocations, which are unacceptable in the
kernel.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>