Don't use kref on nodes
Get rid of the reference counting for in-memory nodes, which is just a
hacky way to avoid accidentally freeing root nodes that are remembered
by the superblock structure. Instead, never delete the root nodes as
part of apfs_free_query() and let the caller decide how to deal with
them.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>