Update ->invalidatepage() for new api
The previous commit tried to fix the build in recent kernels by removing
references to the now obsolete ->invalidatepage() address space
operation. I picked it up at first but it broke my tests pretty badly,
so I have to revert it and find another solution.

It seems that the replacement ->invalide_folio() operation has the same
effect, so use it from now on. This is still a hack that will need to be
reevaluated in the future.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>