Fix build for RHEL 9 kernels
RHEL 9 is using the 5.14 kernel, but they have picked up a patch from
5.15 that broke the api for us:

  https://github.com/linux-apfs/linux-apfs-rw/issues/26

They don't seem to provide a macro to check the exact kernel version in
use, so treat all 5.14 RHEL kernels as 5.15. This will of course break
the build for RHEL users who haven't updated their kernels, but I don't
know of a better way to handle the problem.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>