Adjust behaviour on orphan link creation
Make a few changes on orphan link behaviour to better match the official
implementation. It's not clear to me that this is mandatory, but I want
my fsck to work properly with official images: I've already modified it
to that end, so these changes need to be applied here as well to avoid
potential reports of corruption.

Anyway, the changes to orphan creation are four:

  - Report the inode's link count as 0, that is, don't count the orphan
    link itself.
  - Preserve the original primary name for the inode.
  - Preserve the inode's parent id.
  - Update the volume's file count immediately, instead of waiting until
    the orphan gets purged.

In essence, stop treating orphaned inodes as real files.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>