Add support for taking snapshots
Implement an ioctl that creates a snapshot for the current transaction.
A matching userland tool has already been added to apfsprogs.

Note that, at this point, snapshots won't be handled correctly on later
writes at all. Being able to actually create snapshots for testing is
only the first step in that direction.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>