Fix apfs_node_locate_key() for the omap snapshots
Attempts to create a second snapshot on a same volume fail, with the
driver reporting corruption. The problem is in apfs_node_locate_key():
we assume that all nodes with fixed key/value length are omap nodes, and
so we work with a wrong value for the length of keys in the snapshots
tree.

These problems would never occur if we just read the key/value lengths
from the root node info footer, and maybe kept them in a btree struct
shared by all the nodes. For now though, just hardcode the correct
length value for trees of this type.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>