Never skip sb copy for readwrite containers
If I perform the following actions in order:

  o Mount a volume in readwrite mode
  o Mount one of its snapshots
  o Unmount the volume
  o Unmount the snapshot

I end up with an out-of-date superblock at block zero. The problem is
that both apfs_put_super() and apfs_make_super_copy() check if the last
unmounted superblock is read-only before making the copy, ignoring
changes made by other mounts. This became obvious while working with
snapshots, but it should affect regular volumes as well.

Instead check for the APFS_READWRITE flag in the container superblock,
which should be set iff at least one of the container's mounts was not
read-only.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>