Flush multiple xids from free queues when needed
Aditya Garg has provided me with an image that reports a full internal
pool after only a few operations:

  https://github.com/linux-apfs/linux-apfs-rw/issues/17#issuecomment-1282855494

The problem here is that the image has most of its ip allocated for
blocks that are in the free queue, waiting for their turn to get freed.
Those blocks span a large number of transactions, but my driver only
flushes a single xid (the oldest) after every commit, staying always
very close to the limit, which eventually gets reached.

Another consequence of this is that we get constant commits, because the
free queue size checks in apfs_transaction_need_commit() always fail.
So, to fix those two issues at once, modify apfs_flush_free_queue() to
flush as many transactions as needed for those checks to pass. Of course
always leave the most recent ones, otherwise there would be no point in
having a free queue at all.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>