Support reads of LZVN-compressed files
Use Apple's official LZFSE library[0] to add support for LZVN
compression. Their code for LZFSE compression proper is also dumped with
this patch, but it's not actually rigged up for the driver because I
have no image to test it on.

For the record, the build on Linux 5.15 produces the following objtool
warning:

  lzfse/lzvn_decode_base.o: warning: objtool: lzvn_decode()+0x1b0: sibling call from callable instruction with modified stack frame

I'm not sure what this means, but there seems to be no sibling call in
that position, just a regular goto:

   1b0:   e9 00 00 00 00          jmp    1b5 <lzvn_decode+0x1b5>

I'll ignore it for now, but I guess I should report it if it's a false
positive.

[0]: https://github.com/lzfse/lzfse

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>