Support reads of LZFSE-compressed files
Using the official LZFSE library (which I already dumped here in the
previous patch), add support for LZFSE compression. For testing I used
images created with the following tool:

  https://github.com/Siguza/fscmp

I've tested both resource and attribute compression. For resource
compression, I also tested a file that had a single uncompressed block.
That block had 0xff in the first byte, but I don't know if that's a
general rule so it's better to just check for the first byte of the
LZFSE magic.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>