Fix cknodes mount option on readwrite mode
Luflosi has reported a long time ago that the cknodes mount option does
not work well on readwrite mounts:

  https://github.com/linux-apfs/linux-apfs-rw/issues/15

I don't care much for the cknodes option, and I don't know if anybody is
actually using it, but I shouldn't ignore a simple bug like this for
such a long time.

The problem here is that there are several places in the code where an
object will get its checksum verified after getting changed, but before
the transaction is committed and the checksum updated. As a solution,
refuse to verify buffers that are already part of a transaction. Try to
make sure that a check happens before the join.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>