Avoid double brelse of spaceman on error
I ran into a stacktrace while testing the cknodes option for the
previous patch. The error message was "VFS: brelse: Trying to free free
buffer", and it showed up when apfs_read_node() failed while flushing a
free queue. The problem is in the never tested cleanup code for
apfs_read_spaceman(): the buffer is released, but dangling pointers
remain in the superblock.

At some point I need to start testing with error injection to find stuff
like this, but for now at least fix the bugs that show up during other
work, such as this one.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>