0.3.0
Some packagers have requested that the driver gets tagged releases:

  https://github.com/linux-apfs/linux-apfs-rw/pull/35

So I'll use 0.3.0 at this point, because the dkms file was set to 0.2.

This should not be understood to imply that the driver is in any way
ready. Writes are still very dangerous and a lot of basic features are
completely missing, for example preserving old snapshots.

Read-only support is somewhat more complete, with sealed volumes,
snapshots, and all the missing compression algorithms recently added.
Encryption is still not in the works though.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>